backend/pyproject.toml
[project]
name = "library"
version = "0.1.0"
description = "Neighborhood Library — gRPC backend"
requires-python = ">=3.12,<3.13"
readme = "README.md"

# Runtime dependency set. Phase 1 added the gRPC stack and config plumbing;
# Phase 2 adds persistence (SQLAlchemy 2.0 async + asyncpg + Alembic + the
# greenlet shim that SQLAlchemy's async layer requires under the hood).
# Testcontainers lands in Phase 4 alongside integration tests.
dependencies = [
    "grpcio>=1.66,<2.0",
    "grpcio-tools>=1.66,<2.0",
    "grpcio-health-checking>=1.66,<2.0",
    "grpcio-reflection>=1.66,<2.0",
    "python-dotenv>=1.0,<2.0",
    "pydantic>=2.9,<3.0",
    "pydantic-settings>=2.5,<3.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.29,<1.0",
    "alembic>=1.13,<2.0",
    "greenlet>=3.0,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "testcontainers[postgresql]>=4.8,<5.0",
    "psycopg[binary]>=3.2,<4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/library"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
# Share one event loop across the whole test session so the testcontainer-
# Postgres + in-process gRPC server fixtures (which are session-scoped) can
# stay alive across tests, and so the gRPC stubs/channels created in those
# fixtures aren't bound to a per-test loop that gets torn down between
# tests. Each test still gets a fresh DB state via the autouse TRUNCATE
# fixture in tests/conftest.py.
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
